+++ src/thirds/application/SignUpMenu.java
@@ -1,10 +1,10 @@
 package thirds.application;
 
 import thirds.io.Resources;
-import thirds.scratch.LogInScreen;
-import thirds.scratch.LoggedInScreen;
-import thirds.scratch.SignUpScreen;
-import thirds.scratch.SignUpScrollable;
+import thirds.interfaces.LogInScreen;
+import thirds.interfaces.LoggedInScreen;
+import thirds.interfaces.SignUpScreen;
+import thirds.interfaces.SignUpScrollable;
 import thirds.swing.MoveableComponent;
 
 import javax.imageio.ImageIO;
