+++ src/thirds/interfaces/PanelMaker.java
@@ -1,4 +1,4 @@
-package thirds.scratch;
+package thirds.interfaces;
 
 import thirds.io.Resources;
